Adjust Readline::feed method to accept a byte instead of an int
libreadline, despite accepting an int to the rl_stuff_char function,
actually just treats the value passed in as a char, i.e., a single byte.
This can cause confusion when it comes to Unicode characters. To make
this problem more explicit, this change adjusts the signature of the
Readline::feed method to accept a u8 instead of the previously used
c_int.

+++ examples/basic.rs
@@ -53,7 +53,7 @@ fn process_byte<W>(w: &mut W, rl: &mut Readline, byte: u8) -> Result<Op>
 where
   W: Write,
 {
-  if let Some(text) = rl.feed(i32::from(byte)) {
+  if let Some(text) = rl.feed(byte) {
     // Check whether our `Readline` object has completed a line given
     // the user-provided input. If so, check whether the user typed
     // "quit" and exit. If not just print it, move to the next line,
